Rename processors directory to generators
